move conversation to main folder
