+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/AllConversationScreen.kt
@@ -1,6 +1,5 @@
-package com.wire.android.ui.conversation
+package com.wire.android.ui.main.conversation
 
-import android.util.Log
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -13,14 +12,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.wire.android.R
 import com.wire.android.ui.common.UserProfileAvatar
-import com.wire.android.ui.conversation.all.model.Conversation
-import com.wire.android.ui.conversation.all.model.ConversationFolder
-import com.wire.android.ui.conversation.all.model.NewActivity
-import com.wire.android.ui.conversation.all.model.toUserInfoLabel
-import com.wire.android.ui.conversation.common.EventBadgeFactory
-import com.wire.android.ui.conversation.common.FolderHeader
-import com.wire.android.ui.conversation.common.RowItem
-import com.wire.android.ui.conversation.common.UserLabel
+import com.wire.android.ui.main.conversation.all.model.Conversation
+import com.wire.android.ui.main.conversation.all.model.ConversationFolder
+import com.wire.android.ui.main.conversation.all.model.NewActivity
+import com.wire.android.ui.main.conversation.all.model.toUserInfoLabel
+import com.wire.android.ui.main.conversation.common.EventBadgeFactory
+import com.wire.android.ui.main.conversation.common.FolderHeader
+import com.wire.android.ui.main.conversation.common.RowItem
+import com.wire.android.ui.main.conversation.common.UserLabel
 
 
 @Composable
