+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/ConversationRoute.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.conversation
+package com.wire.android.ui.main.conversation
 
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
@@ -12,7 +12,7 @@ import com.wire.android.R
 import com.wire.android.ui.common.FloatingActionButton
 import com.wire.android.ui.common.WireBottomNavigationBar
 import com.wire.android.ui.common.WireBottomNavigationItemData
-import com.wire.android.ui.conversation.navigation.ConversationsNavigationItem
+import com.wire.android.ui.main.conversation.navigation.ConversationsNavigationItem
 
 
 @Composable
