+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/common/FolderHeader.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.conversation.common
+package com.wire.android.ui.main.conversation.common
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.MaterialTheme
