+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/common/RowItem.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.conversation.common
+package com.wire.android.ui.main.conversation.common
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
