+++ app/src/main/kotlin/com/wire/android/ui/main/conversation/mock/Mock.kt
@@ -1,20 +1,20 @@
-package com.wire.android.ui.conversation
+package com.wire.android.ui.main.conversation
 
-import com.wire.android.ui.conversation.all.model.AvailabilityStatus
-import com.wire.android.ui.conversation.all.model.Conversation
-import com.wire.android.ui.conversation.all.model.ConversationFolder
-import com.wire.android.ui.conversation.all.model.ConversationInfo
-import com.wire.android.ui.conversation.all.model.EventType
-import com.wire.android.ui.conversation.all.model.Membership
-import com.wire.android.ui.conversation.all.model.NewActivity
-import com.wire.android.ui.conversation.all.model.UserInfo
-import com.wire.android.ui.conversation.call.model.Call
-import com.wire.android.ui.conversation.call.model.CallEvent
-import com.wire.android.ui.conversation.call.model.CallInfo
-import com.wire.android.ui.conversation.call.model.CallTime
-import com.wire.android.ui.conversation.mention.model.Mention
-import com.wire.android.ui.conversation.mention.model.MentionInfo
-import com.wire.android.ui.conversation.mention.model.MentionMessage
+import com.wire.android.ui.main.conversation.all.model.AvailabilityStatus
+import com.wire.android.ui.main.conversation.all.model.Conversation
+import com.wire.android.ui.main.conversation.all.model.ConversationFolder
+import com.wire.android.ui.main.conversation.all.model.ConversationInfo
+import com.wire.android.ui.main.conversation.all.model.EventType
+import com.wire.android.ui.main.conversation.all.model.Membership
+import com.wire.android.ui.main.conversation.all.model.NewActivity
+import com.wire.android.ui.main.conversation.all.model.UserInfo
+import com.wire.android.ui.main.conversation.call.model.Call
+import com.wire.android.ui.main.conversation.call.model.CallEvent
+import com.wire.android.ui.main.conversation.call.model.CallInfo
+import com.wire.android.ui.main.conversation.call.model.CallTime
+import com.wire.android.ui.main.conversation.mention.model.Mention
+import com.wire.android.ui.main.conversation.mention.model.MentionInfo
+import com.wire.android.ui.main.conversation.mention.model.MentionMessage
 
 val mockConversations = listOf(
     Conversation(
